# backend/openapi.json
{
  "components" : {
    "schemas" : {
      "AllFlags" : {
        "properties" : {
          "flags" : {
            "additionalProperties" : {
              "type" : "boolean"
            },
            "type" : "object"
          },
          "revision" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "flags", "revision" ],
        "type" : "object"
      },
      "AuditEventView" : {
        "properties" : {
          "action" : {
            "type" : "string"
          },
          "actor" : {
            "type" : "string"
          },
          "details" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "occurredAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "targetKey" : {
            "type" : "string"
          }
        },
        "required" : [ "action", "actor", "id", "occurredAt", "targetKey" ],
        "type" : "object"
      },
...
      "GroupFlags" : {
        "properties" : {
          "flags" : {
            "additionalProperties" : {
              "type" : "boolean"
            },
            "type" : "object"
          },
          "group" : {
            "type" : "string"
          },
          "revision" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "flags", "group", "revision" ],
        "type" : "object"
      },
...
      "LoginRequest" : {
        "properties" : {
          "password" : {
            "minLength" : 1,
            "type" : "string"
          },
          "username" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "password", "username" ],
        "type" : "object"
      },
      "OneFlag" : {
        "properties" : {
          "enabled" : {
            "type" : "boolean"
          },
          "key" : {
            "type" : "string"
          }
        },
        "required" : [ "enabled", "key" ],
        "type" : "object"
      },
      "PageMetadata" : {
        "properties" : {
          "number" : {
            "format" : "int64",
            "type" : "integer"
          },
          "size" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "number", "size", "totalElements", "totalPages" ],
        "type" : "object"
      },
      "PagedModelAuditEventView" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/AuditEventView"
            },
            "type" : "array"
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMetadata"
          }
        },
        "required" : [ "content", "page" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/admin/audit" : {
      "get" : {
        "operationId" : "list_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "targetKey",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedModelAuditEventView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "audit-controller" ]
      }
    },
...
    "/api/v1/auth/login" : {
      "post" : {
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
...
    "/api/v1/evaluate/flags" : {
      "get" : {
        "operationId" : "all",
        "parameters" : [ {
          "in" : "header",
          "name" : "If-None-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AllFlags"
                }
              }
            },
            "description" : "Current values",
            "headers" : {
              "Cache-Control" : {
                "description" : "no-cache",
                "style" : "simple"
              },
              "ETag" : {
                "description" : "\"<revision>\"",
                "style" : "simple"
              }
            }
          },
          "304" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AllFlags"
                }
              }
            },
            "description" : "Not modified (If-None-Match matches)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AllFlags"
                }
              }
            },
            "description" : "Missing, expired or invalid token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AllFlags"
                }
              }
            },
            "description" : "Token without flags:read / feature-flag-service"
          }
        },
        "tags" : [ "evaluation-controller" ]
      }
    },
    "/api/v1/evaluate/flags/{groupKey}/{flagKey}" : {
      "get" : {
        "operationId" : "flag",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupKey",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "flagKey",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "If-None-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OneFlag"
                }
              }
            },
            "description" : "Current values",
            "headers" : {
              "Cache-Control" : {
                "description" : "no-cache",
                "style" : "simple"
              },
              "ETag" : {
                "description" : "\"<revision>\"",
                "style" : "simple"
              }
            }
          },
          "304" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OneFlag"
                }
              }
            },
            "description" : "Not modified (If-None-Match matches)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OneFlag"
                }
              }
            },
            "description" : "Missing, expired or invalid token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OneFlag"
                }
              }
            },
            "description" : "Token without flags:read / feature-flag-service"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OneFlag"
                }
              }
            },
            "description" : "Unknown key"
          }
        },
        "tags" : [ "evaluation-controller" ]
      }
    },
    "/api/v1/evaluate/groups/{groupKey}" : {
      "get" : {
        "operationId" : "group",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupKey",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "If-None-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupFlags"
                }
              }
            },
            "description" : "Current values",
            "headers" : {
              "Cache-Control" : {
                "description" : "no-cache",
                "style" : "simple"
              },
              "ETag" : {
                "description" : "\"<revision>\"",
                "style" : "simple"
              }
            }
          },
          "304" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupFlags"
                }
              }
            },
            "description" : "Not modified (If-None-Match matches)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupFlags"
                }
              }
            },
            "description" : "Missing, expired or invalid token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupFlags"
                }
              }
            },
            "description" : "Token without flags:read / feature-flag-service"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupFlags"
                }
              }
            },
            "description" : "Unknown key"
          }
        },
        "tags" : [ "evaluation-controller" ]
      }
    }
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}